Refactor `StepStateIcon` to use new tooltip component
